Day 9

+++ lib/vector.h
@@ -71,28 +71,29 @@ int cv_get_pos(CoordinateVector* cv, coordinate c)
 }
 
 typedef struct {
-    int32_t* elements;
-    int32_t capacity;
-    int32_t size;
+    int64_t* elements;
+    int64_t capacity;
+    int64_t size;
 } IntVector;
 
+// Should be changed to iv_create_vector
 IntVector* iv_create_coordinate_vector(int capacity)
 {
     IntVector *created = malloc(sizeof *created);
-    created->elements = malloc((sizeof(uint32_t)) * capacity);
+    created->elements = malloc((sizeof(int64_t)) * capacity);
     created->capacity = capacity;
     created->size = 0;
     return created;
 }
 
-void iv_add_int(IntVector* iv, int32_t to_add)
+void iv_add_int(IntVector* iv, int64_t to_add)
 {
     iv->elements[iv->size++] = to_add;
 
     if (iv->capacity == iv->size) // realloc in this case
     { 
-        int32_t capacity_new = iv->capacity * iv->capacity;
-        int32_t* aux = malloc((sizeof(uint32_t)) * capacity_new);
+        int64_t capacity_new = iv->capacity * iv->capacity;
+        int64_t* aux = malloc((sizeof(int64_t)) * capacity_new);
 
         for(int i = 0; i < iv->capacity; ++i)
         {
